refactor(agents): extract AgentActivityStream into shared component
Moved ActivityTab + all helpers (~350 lines) from AgentDetail.tsx
into web/src/components/shared/AgentActivityStream.tsx:
- buildTimeline, humanizeEvent, eventIcon, matchesFilter
- TimelineEvent, EventFilter, FILTER_LABELS types
- SSE connection + activity fetch logic

AgentDetail now composes: <AgentActivityStream agentName=... />
